updated user validations
