+++ src/App.jsx
@@ -8,7 +8,7 @@ import {Chart} from 'react-chartjs-2';
 import theme from './assets/muiTheme';
 import GlobalStyles from './assets/GlobalStyles';
 import {draw} from './mixins/chartjs';
-import Routes from './Routes/Routes';
+import Routes from './Routes';
 import LocalizedDateFnsUtils from "./mixins/LocalizedDateFnsUtils";
 import {MainLayout, MinimalLayout} from "./layouts";
 import {Notification, Spinner} from "./components";
@@ -19,7 +19,6 @@ Chart.helpers.extend(Chart.elements.Rectangle.prototype, {draw: draw});
 const App = () => {
     const history = useHistory();
     const isAuthenticated = useSelector(store => store.auth.isAuthenticated);
-    // const isAuthenticated = "true"
     const isLoading = useSelector(store => store.auth.loading);
     const notification = useSelector(store => store.config.notification);
     const dispatch = useDispatch();
