+++ src/pages/Home/index.jsx
@@ -5,17 +5,18 @@ import ProjectAdminView from "./ProjectAdminView";
 import SiteOwnerView from "./SiteOwnerView";
 import MemberView from "./MemberView";
 const Home = ({ history, location }) => {
-    const user = localStorage.getItem("user");
+    const role = localStorage.getItem("user");
     const project_id = localStorage.getItem('project_id');
+    
     if (project_id) {
         return (
             <Container title="Home">
-            {user==="site_owner" &&   //site-owner and admin view are currently same till we add site-owner controls
-            <SiteOwnerView history={history}></SiteOwnerView>}      
-            {user==="project_admin" &&              
-             <ProjectAdminView history={history}></ProjectAdminView>}
-             {user==="project_member" && 
-             <MemberView history={history}></MemberView>}
+                {role==="site_owner" &&   //site-owner and admin view are currently same till we add site-owner controls
+                    <SiteOwnerView  history={history}></SiteOwnerView>}      
+                {role==="project_admin" &&              
+                    <ProjectAdminView history={history}></ProjectAdminView>}
+            <hr/>
+                    <MemberView history={history}></MemberView>
             </Container>
         );
     }
